Extract a base Protocol class from Gmp
This class provides the base methods for communication over a GVM
protocol.
